requirements.txt
apscheduler==3.10.4
httpx==0.27.2
lxml==5.3.0
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1
streamlit-autorefresh==1.0.1
streamlit==1.39.0
pytest==8.3.3
requests==2.32.3
